requirements.txt
mcp[cli]>=1.10.0
pyyaml>=6.0.1
